+++ src/transformations/buildColumn.ts
@@ -4,7 +4,7 @@ import {
   getContextAndDataSet,
 } from "../utils/codapPhone/index";
 import { reportTypeErrorsForRecords } from "./util";
-import { DDTransformationState } from "../transformation-components/DDTransformation";
+import { DDTransformationState } from "../transformation-components/DataDrivenTransformation";
 import { readableName } from "../transformation-components/util";
 
 /**
