+++ src/transformations/compare.ts
@@ -4,7 +4,7 @@ import { diffArrays } from "diff";
 import { intersectionWithPredicate, unionWithPredicate } from "../utils/sets";
 import { uncheckedFlatten } from "./flatten";
 import { eraseFormulas, getAttributeDataFromDataset } from "./util";
-import { DDTransformationState } from "../transformation-components/DDTransformation";
+import { DDTransformationState } from "../transformation-components/DataDrivenTransformation";
 import { getContextAndDataSet } from "../utils/codapPhone";
 import { readableName } from "../transformation-components/util";
