+++ src/transformations/join.ts
@@ -1,7 +1,7 @@
 import { CodapAttribute, Collection } from "../utils/codapPhone/types";
 import { DataSet } from "./types";
 import { uniqueName } from "../utils/names";
-import { DDTransformationState } from "../transformation-components/DDTransformation";
+import { DDTransformationState } from "../transformation-components/DataDrivenTransformation";
 import { getContextAndDataSet } from "../utils/codapPhone";
 import { readableName } from "../transformation-components/util";
